vc1parser: re-enable BDU parsing for bitstream
Currently the parser is not able to parse bitstream like in this pipeline:

  gst-launch-1.0 filesrc location=mytestfile.vc1 ! vc1parse ! ..

Since the parser bails out with an error message.

This patch proposes a way re-enable it using the function
gst_vc1_identify_next_bdu() early. If a complete sequence header BDU
is found, the stream format is assumed as bdu-frame.

https://bugzilla.gnome.org/show_bug.cgi?id=743948